feat: split-factory-service-named-factory-editable-definition-cluster-003 - Preserve session-scoped editable-definition compatibility and remove the legacy factory.go size exception
